calcLocation promise
